27/1added

+++ HyperMarket/Customer Bill.cs
@@ -9,13 +9,13 @@ namespace HyperMarket
     internal class Customer_Bill
     { 
         static int NumberOfBills = 0;
-        public int BillId {get{return BillId;} private set{BillId = NumberOfBills;}}
+        public int BillId { get{return BillId;} private set{BillId = NumberOfBills;}}
         public int CustomerId { get; set; }  
         public String CustomerName { get; set; }
         public int CashierId { get; set; }
         public string CashierName { get; set; }
         public List<ProductNeed> Customer_Product= new List<ProductNeed>();
-        public decimal TotalPrice { get; set; }
+        public double TotalPrice { get; set; }
         public DateTime CreatedTime ;
         public Customer_Bill()
         {
